+++ src/abstraction/eq4Transformateur2/Transformateur2_gestion_stocks.java
@@ -93,62 +93,5 @@ public class Transformateur2_gestion_stocks extends Transformateur2_contratCadre
 			return cout ;
 		}
 		
-	//  Nombre de tours d'autonmies restants pour couvrir les echeanciers de nos contrats cadres	
-		public int nbToursAutonomiePate(PateInterne pate) {
-			int resu = -1;
-			double pate_echeancier = 0;
-			int step = Filiere.LA_FILIERE.getEtape();
-			List<ExemplaireContratCadre> contratsDeCeType = super.getContratDeCeType(pate);
-			while (pate_echeancier < super.getStockPateValeur(pate)) {
-				resu++;
-				double qte_tour = 0;
-				for (ExemplaireContratCadre ex : contratsDeCeType) {
-					qte_tour += super.getQuantiteALivrerAuStep(ex, step + resu);
-				}
-				if (qte_tour == 0) {
-					return 1000;			// On a assez pour subvenir à toutes le echeances
-				}
-				else {
-					pate_echeancier += qte_tour;
-				}
-			}
-			return resu;
-		}
-		
-		public int nbToursAutonomiePateEtFeves(PateInterne pate) {
-			int resu = -1;
-			double pate_echeancier = 0;
-			int step = Filiere.LA_FILIERE.getEtape(); //RETIRE POUR TEST
-			//int step = 1; //pour test
-			List<ExemplaireContratCadre> contratsDeCeType = super.getContratDeCeType(pate);
-			while (pate_echeancier < super.getStockPateValeur(pate)) {
-				resu++;
-				double qte_tour = 0;
-				for (ExemplaireContratCadre ex : contratsDeCeType) {
-					qte_tour += super.getQuantiteALivrerAuStep(ex, step + resu);
-				}
-				if (qte_tour == 0) {
-					return 1000;			// On a assez pour subvenir à toutes le echeances
-				}
-				else {
-					pate_echeancier += qte_tour;
-				}
-			}
-			double equivalent_feves = (pate_echeancier - super.getStockPateValeur(pate))/super.getCoeffTFEP();
-			Feve feve = super.creerFeve(pate);
-			while (equivalent_feves < super.getStockFevesValeur(feve)) {
-				resu++;
-				double qte_tour = 0;
-				for (ExemplaireContratCadre ex : contratsDeCeType) {
-					qte_tour += super.getQuantiteALivrerAuStep(ex, step + resu);
-				}
-				if (qte_tour == 0) {
-					return 1000;			// On a assez pour subvenir à toutes le echeances
-				}
-				else {
-					equivalent_feves += qte_tour/super.getCoeffTFEP();
-				}
-			}
-			return resu;
-		}
+	
 }
